requirements.txt
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.34.74
botocore==1.34.162
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==7.1.2
comm==0.2.2
contourpy==1.3.0
cupy-cuda11x==13.4.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
diskcache==4.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
fire==0.7.0
fonttools==4.58.0
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.7.0
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.2
notebook_shim==0.2.4
numpy==1.26.4
nuscenes-devkit==1.1.9
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.3.8
plyfile==1.0.3
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.20.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.0
s3transfer==0.10.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.13.1
Send2Trash==1.8.3
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tenacity==6.0.0
tensorboardX==2.6.2.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xarray==2022.6.0
yacs==0.1.8
zipp==3.21.0
